refactor(dashboard page_components.py): split code page
split into modules so that maintainability is improved

improve logging by adding exception logs
